requirements.txt
timm
torch
torchvision
hydra-core
wandb
omegaconf